requirements.txt
flask
Pillow
transformers
torch
gunicorn
duckduckgo-search